Use BaseTemplateToolbox hook so link appears in the right place
The SkinTemplateToolboxEnd hook is nasty and requires echo-ing raw HTML.
It also has the problem that if there are any other extensions using
that hook, the CiteThisPage link will easily end up in the wrong place
and under the wrong heading.

This is trivial to fix though by using the BaseTemplateToolbox hook,
which is designed for actually adding links to the toolbox!

Bug: T140290
Change-Id: I9557130c40e17edc12a225a3b7b1cdec028f93d0