Add a possibility to add a "cite this page" link without adding the NS as a content namespace
Sometimes it's required or at least a good option to have a cite this page
link on pages, that aren't content pages (like the project namespace).

This change adds a new configuration option for this extensions to allow to
add namespaces as namespaces that should have a cite this page link, even if
they're not a content page.

Bug: T150053
Change-Id: I2523bd9124ab8998a7c327d8045d7585d4dcf8b4