Fix dangling comma-space-dot on Special:CiteThisPage
Examples of this issue in production:

* https://nl.wikipedia.org/wiki/Maan -> "Cite ethis page" ->
  https://nl.wikipedia.org/w/index.php?title=Speciaal:Citeren&page=Maan&id=64037828&wpFormIdentifier=titleform
  > Uitgever: Wikipedia, .

* https://es.wikipedia.org/wiki/Luna -> "Cite this page" ->
  https://es.wikipedia.org/w/index.php?title=Especial:Citar&page=Luna&id=150096436&wpFormIdentifier=titleform
  > Publisher: Wikipedia, .

* https://www.mediawiki.org/wiki/MediaWiki:Citethispage-content
  > Publisher: MediaWiki, .

The "sitesubtitle" message has been empty by default for nearly
twenty years (since 2005, T6192) and is tolerated as such by the Skin.

The CiteThisPage extension as such should not unconditionally render
it in a way that relies on it having a non-empty value.

Credit to Nicole Sharp for raising this issue at
<https://www.mediawiki.org/wiki/MediaWiki_talk:Citethispage-content>.

I considered replacing it with {{SERVERNAME}}, but decided to instead
remove it since the domain name is already given in the permalink
on the same line.

Change-Id: If39a2b6dd4dda7309f538a40a1d567b479f99ce8